refacotr(vsmgr: cli): move 'util sealer sectors' into a single file
